vdk-test-utils: Apply changes after fully managed cursor merge
Fully managed cursor implementation required extending vdk-test-utils.
Due to present modules testing and release lifecycles, changes of
interdependent packages need to be made rather sequentially, in a
backwards-compatible fashion - for the pipeline to complete.
A decision was made for the classes changed to be placed within
vdk-core test itself, then moved to the test suite afterwards.

Moved test_run_sql_queries.py helper classes to the dedicated
util_plugins.py. Cleaned up the todos.

Testing Done: did run locally vdk-core and vdk-test-utils tests
